19341 FIX Fix dashboard crashes after update due to incompatible graph size type
The graph size type was changed from float to
int by accident (commit e953050805b2a0336714903fd8ae509f981cd550).

With this change we set the type back to float.
This allows old data to be used as well as newer data (int to float
conversion is implicit as it does not incur data loss).

Jira: SUP-27803

+++ cmk/gui/graphing/_utils.py
@@ -17,7 +17,7 @@ class MKCombinedGraphLimitExceededError(MKHTTPException):
     status = http.HTTPStatus.BAD_REQUEST
 
 
-SizeEx = NewType("SizeEx", int)
+SizeEx = NewType("SizeEx", float)
 
 
 def get_graph_data_from_livestatus(only_sites, host_name, service_description):
